Remove supporting composer dependencies
This commit removes dependencies that are required to make Composer work
properly with Drupal. These dependencies are also included in the Open
Social repository so duplicating them doesn't make that much sense.

The configuration for these modules is still in the template but the
versions are now upgrade with Open Social itself.

The change in `extra` and `scripts` configuration prepares for Open
Social's switch to `drupal/core-composer-scaffold`.