Fix composer scaffold for Open Social 9.0
The change to the scaffolding plugin is introduced in 10.0 of Open
Social as it requires people to make changes to their composer.json
(thus a breaking change). The new scaffolding plugin no longer requires
a top-level script to be added as it adds a composer plugin. However,
the old scaffolding plugin still requires these scripts. Thus we must
add them to our composer.json until we start installing Open Social 10.0
and newer.